serving-fastapi/pyproject.toml
[project]
name = "voltage-hub-serving-fastapi"
version = "0.1.0"
description = "Voltage Hub serving API"
requires-python = ">=3.11"
dependencies = [
    "cachetools>=5.5.2",
    "fastapi>=0.115.12",
    "google-cloud-bigquery>=3.40.1",
    "pydantic>=2.11.3",
    "uvicorn>=0.34.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.3.5",
]

[tool.pytest.ini_options]
pythonpath = ["."]
